requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0
django-bootstrap-datepicker-plus==4.0.0
django-colorfield==0.6.3
django-countries==7.2.1
django-language-field==0.0.3
django-tz-detect==0.3.0
idna==3.3
ipaddress==1.0.16
numpy==1.22.0
opencv-python==4.5.5.62
pdf2image==1.16.0
Pillow==8.4.0
pygeoip==0.3.2
PyPDF2==1.26.0
python-decouple==3.5
pytz==2016.6
razorpay==1.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
